# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Resonance API Documentation",
        "description": "API documentation for Resonance",
        "contact": {
            "email": "casier07@hotmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://resonance-be.ddev.site",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/locations/{id}": {
            "get": {
                "tags": [
                    "Locations"
                ],
                "summary": "Get a specific location",
                "operationId": "bcaac2b6f85e83cdc99139be535c66e6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Location UUID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Location details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Location"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Location not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Locations"
                ],
                "summary": "Update a location",
                "operationId": "3ac5d4120b6fb23f55f179649493a2d8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Location UUID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Sportpaleis Antwerpen"
                                    },
                                    "street": {
                                        "type": "string",
                                        "example": "Schijnpoortweg",
                                        "nullable": true
                                    },
                                    "house_number": {
                                        "type": "string",
                                        "example": "119",
                                        "nullable": true
                                    },
                                    "city": {
                                        "type": "string",
                                        "example": "Antwerpen",
                                        "nullable": true
                                    },
                                    "postal_code": {
                                        "type": "string",
                                        "example": "2170",
                                        "nullable": true
                                    },
                                    "country": {
                                        "type": "string",
                                        "example": "Belgium"
                                    },
                                    "latitude": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 51.2306,
                                        "nullable": true
                                    },
                                    "longitude": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 4.4452,
                                        "nullable": true
                                    },
                                    "website": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "description": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "image_url": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "source": {
                                        "type": "string",
                                        "enum": [
                                            "manual",
                                            "api"
                                        ]
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "pending_approval",
                                            "verified",
                                            "rejected"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Location updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Location"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Location not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Locations"
                ],
                "summary": "Delete a location",
                "operationId": "721591ee086203214e6eb54f25625656",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Location UUID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Location deleted successfully"
                    },
                    "404": {
                        "description": "Location not found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Location": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "123e4567-e89b-12d3-a456-426614174000"
                    },
                    "name": {
                        "type": "string",
                        "example": "Sportpaleis Antwerpen"
                    },
                    "street": {
                        "type": "string",
                        "example": "Schijnpoortweg",
                        "nullable": true
                    },
                    "house_number": {
                        "type": "string",
                        "example": "119",
                        "nullable": true
                    },
                    "city": {
                        "type": "string",
                        "example": "Antwerpen",
                        "nullable": true
                    },
                    "postal_code": {
                        "type": "string",
                        "example": "2170",
                        "nullable": true
                    },
                    "country": {
                        "type": "string",
                        "example": "Belgium"
                    },
                    "latitude": {
                        "type": "number",
                        "format": "float",
                        "example": 51.2306,
                        "nullable": true
                    },
                    "longitude": {
                        "type": "number",
                        "format": "float",
                        "example": 4.4452,
                        "nullable": true
                    },
                    "website": {
                        "type": "string",
                        "nullable": true
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "image_url": {
                        "type": "string",
                        "nullable": true
                    },
                    "source": {
                        "type": "string",
                        "example": "manual"
                    },
                    "status": {
                        "type": "string",
                        "example": "verified"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        }
    },
...
}